requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-multipart==0.0.9
watchdog==4.0.0
PyPDF2==3.0.0
beautifulsoup4==4.12.3
requests==2.31.0
sentence-transformers==3.0.0
faiss-cpu
numpy
langchain-community==0.2.0